+++ workbench/database/migrations/2026_01_01_000002_create_onboarding_tour_steps_table.php
@@ -1,0 +1,3 @@
+<?php
+
+return require __DIR__ . '/../../../database/migrations/create_onboarding_tour_steps_table.php.stub';