+++ workbench/database/migrations/2026_01_01_000005_add_translations_to_onboarding_tables.php
@@ -1,0 +1,3 @@
+<?php
+
+return require __DIR__ . '/../../../database/migrations/add_translations_to_onboarding_tables.php.stub';