+++ workbench/database/migrations/2026_01_01_000006_add_meta_to_onboarding_tour_completions_table.php
@@ -1,0 +1,3 @@
+<?php
+
+return require __DIR__ . '/../../../database/migrations/add_meta_to_onboarding_tour_completions_table.php.stub';